+++ src/services/games.service.ts
@@ -123,10 +123,8 @@ export class GamesService {
         `Early Access: ${foundGame.early_access} -> ${game.early_access}`,
       );
     }
-    if (foundGame.direct_play !== game.direct_play) {
-      differences.push(
-        `Early Access: ${foundGame.direct_play} -> ${game.direct_play}`,
-      );
+    if (foundGame.type !== game.type) {
+      differences.push(`Game Type: ${foundGame.type} -> ${game.type}`);
     }
     if (foundGame.version != game.version) {
       differences.push(`Version: ${foundGame.version} -> ${game.version}`);